# public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Butwal Hacks Internal API",
    "description": "Complete API reference for the Butwal Hacks platform. Covers all 18 internal API routes.\n\n**Base URL:** `http://localhost:3000` (development) or `https://butwalhacks.com` (production)\n\n**Auth:** Most endpoints require a valid Supabase session. The session token is managed via cookies.\n\n**Methods:** All routes use standard Next.js API route handlers.",
    "version": "1.0.0",
    "contact": {
      "name": "Butwal Hacks Maintainers",
      "url": "https://butwalhacks.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://butwalhacks.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__session",
        "description": "Auth0 session cookie. Set automatically after login via `auth()` from `@auth0/nextjs-auth0`. Most endpoints extract the user ID from the active session."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token for development/testing. All auth is session-based via Auth0."
      }
    },
    "schemas": {
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" }
        }
      }
    }
  },
  "security": [
    { "sessionAuth": [] }
  ],
  "paths": {
...
    "/api/webhooks/auth0": {
      "post": {
        "summary": "Auth0 webhook receiver",
        "description": "Receives webhook events from Auth0 for user, organization, and membership sync. Verifies using Svix signatures.\n\nprofile sync during login/signup** `user.created`, `user.updated`, `organization.*`, `organizationMembership.*`",
        "tags": ["Webhooks"],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": { "type": "string", "description": "Auth0 event type" },
                  "data": { "type": "object", "description": "Event payload (varies by type)" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Event processed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "400": { "description": "Missing Svix headers or verification failed" },
          "500": { "description": "Database error" }
        }
      }
    },
...
    "/api/projects": {
      "post": {
        "summary": "Submit a new project",
        "description": "Creates a new project entry. Auto-sets `profile_id` from the authenticated user's session.",
        "tags": ["Projects"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": { "type": "string", "description": "Project title" },
                  "description": { "type": "string", "nullable": true },
                  "demo_url": { "type": "string", "format": "uri", "nullable": true },
                  "github_url": { "type": "string", "format": "uri", "nullable": true },
                  "cover_image": { "type": "string", "format": "uri", "nullable": true },
                  "video_url": { "type": "string", "format": "uri", "nullable": true },
                  "tech_stack": { "type": "array", "items": { "type": "string" } },
                  "team_id": { "type": "string", "format": "uuid", "nullable": true },
                  "event_id": { "type": "string", "format": "uuid", "nullable": true }
                }
              },
              "example": {
                "title": "AgriTech Smart Monitor",
                "description": "IoT-based crop monitoring system",
                "tech_stack": ["React", "Python", "IoT"],
                "github_url": "https://github.com/user/agritech"
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Project created" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal server error" }
        }
      }
    },
...
    "/api/reviews": {
      "post": {
        "summary": "Submit event review",
        "description": "Submits or updates a star rating for an event. Uses upsert so re-rating updates the existing entry.",
        "tags": ["Events"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["event_id", "rating"],
                "properties": {
                  "event_id": { "type": "string", "format": "uuid" },
                  "rating": { "type": "integer", "minimum": 1, "maximum": 5, "description": "Star rating (1-5)" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Review submitted" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal server error" }
        }
      }
    },
...
    "/api/resources/complete": {
      "post": {
        "summary": "Mark resource as completed",
        "description": "Marks a learning resource as completed by the authenticated user.",
        "tags": ["Resources"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["resource_id"],
                "properties": {
                  "resource_id": { "type": "string", "format": "uuid", "description": "Resource ID to mark complete" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Resource marked complete" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/api/teams": {
      "post": {
        "summary": "Create a team",
        "description": "Creates a new team and automatically adds the creator as captain. Sets `is_captain: true` for the creator in `team_members`.",
        "tags": ["Teams"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "event_id"],
                "properties": {
                  "name": { "type": "string", "description": "Team name" },
                  "event_id": { "type": "string", "format": "uuid", "description": "Event the team is for" }
                }
              },
              "example": {
                "name": "Team Phoenix",
                "event_id": "abc-123-456"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Team created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "team": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string", "format": "uuid" },
                        "name": { "type": "string" },
                        "event_id": { "type": "string", "format": "uuid" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal server error" }
        }
      }
    },
...
  },
...
}